+++ apps/nextjs/app-router/src/contexts/auth.tsx
@@ -1,5 +1,5 @@
 "use client"
-import { AuthProvider as AuraAuthProvider, type AuthProviderProps } from "@aura-stack/next"
+import { AuthProvider as AuraAuthProvider, type AuthProviderProps } from "@aura-stack/next/client"
 import { authClient } from "@/lib/auth-client"
 
 export const AuthProvider = ({ children, initialSession }: Omit<AuthProviderProps, "client">) => {
