+++ apps/nextjs/app-router/src/lib/auth.ts
@@ -1,7 +1,11 @@
-import { createAuth, builtInOAuthProviders, type BuiltInOAuthProvider } from "@aura-stack/next"
+import { createAuth } from "@aura-stack/next"
+//import { builtInOAuthProviders, type BuiltInOAuthProvider } from "@aura-stack/next/oauth"
+//console.log("Built-in OAuth providers:", builtInOAuthProviders)
 
-export const oauth = Object.keys(builtInOAuthProviders) as BuiltInOAuthProvider[]
-export const providers = [builtInOAuthProviders.github(), builtInOAuthProviders.gitlab(), builtInOAuthProviders.bitbucket()]
+//export const oauth = Object.keys(builtInOAuthProviders) as BuiltInOAuthProvider[]
+export const oauth = []
+//export const providers = [builtInOAuthProviders.github(), builtInOAuthProviders.gitlab(), builtInOAuthProviders.bitbucket()]
+export const providers = [{}, {}, {}]
 
 export const { api, core } = createAuth({
     oauth,
