+++ packages/core/src/actions/signIn/authorization-url.ts
@@ -1,7 +1,7 @@
 import { OAuthProvider } from "@/@types/index.ts"
 import { AuthInternalError } from "@/shared/errors.ts"
 import { OAuthAuthorization } from "@/schemas.ts"
-import { createPKCE, createSecretValue } from "@/shared/security.ts"
+import { createPKCE, createSecretValue } from "@/shared/crypto.ts"
 import type { GlobalContext } from "@aura-stack/router"
 
 export const setSearchParams = (url: URL, params: Record<string, string | undefined>) => {
