+++ packages/core/src/client/index.ts
@@ -1,5 +1,1 @@
-export { createClient, createAuthClient } from "@/client/client.ts"
-
-export type { AuthClient, AuthClientOptions, SignInOptions, SignOutOptions } from "@/@types/index.ts"
-
-export type { Client, ClientOptions } from "@aura-stack/router/types"
+export { createAuthClient, type AuthClientOptions } from "@/client/client.ts"
