+++ packages/core/src/shared/crypto.ts
