+++ packages/core/src/shared/identity.ts
@@ -1,5 +1,7 @@
 import { string, z } from "zod/v4"
-import { EditableShape } from "@/@types/utility.ts"
+import type { EditableShape } from "@/@types/utility.ts"
+
+export type { InferAuthIdentity, InferIdentity, InferShape, ShapeToObject, EditableShape } from "@/@types/utility.ts"
 
 export const UserIdentity = z.object({
     sub: string(),
@@ -7,8 +9,6 @@ export const UserIdentity = z.object({
     image: string().nullable().optional(),
     email: string().nullable().optional(),
 })
-
-export const StrippedUserIdentity = UserIdentity.omit({ sub: true })
 
 export type UserShape = (typeof UserIdentity)["shape"]
 export type UserIdentityType = z.infer<typeof UserIdentity>
