+++ packages/core/src/shared/index.ts
@@ -1,0 +1,5 @@
+/**
+ * This file re-exports the public functions
+ */
+export { createBasicAuthHeader } from "@/shared/utils.ts"
+export { createSyslogMessage } from "@/shared/logger.ts"
