+++ packages/core/test/secure.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect } from "vitest"
-import { createPKCE } from "@/shared/security.ts"
+import { createPKCE } from "@/shared/crypto.ts"
 
 describe("createPKCE", () => {
     test("generates a valid code verifier and code challenge", async () => {
