+++ packages/next/deno.json
@@ -5,7 +5,11 @@
   "exports": {
     ".": "./src/index.ts",
     "./client": "./src/client.ts",
-    "./oauth": "./src/oauth/index.ts"
+    "./oauth": "./src/oauth/index.ts",
+    "./identity": "./src/_core/identity.ts",
+    "./crypto": "./src/_core/crypto.ts",
+    "./shared": "./src/_core/shared.ts",
+    "./types": "./src/@types/index.ts"
   },
   "imports": {
     "@/": "./src/",
