+++ packages/next/src/_core/shared.ts
@@ -1,0 +1,1 @@
+export * from "@aura-stack/react/shared"
