+++ packages/next/src/index.ts
@@ -1,4 +1,2 @@
 export { createAuth } from "@/createAuth"
-export * from "@/client"
-export { builtInOAuthProviders, type BuiltInOAuthProvider } from "@aura-stack/auth/oauth/index"
 export type { User, Session, AuthConfig, AuthInstance } from "@/@types/core.ts"
