+++ packages/react/package.json
@@ -48,6 +48,21 @@
       "import": "./dist/oauth/index.js",
       "require": "./dist/oauth/index.cjs"
     },
+    "./identity": {
+      "types": "./dist/_core/identity.d.ts",
+      "import": "./dist/_core/identity.js",
+      "require": "./dist/_core/identity.cjs"
+    },
+    "./crypto": {
+      "types": "./dist/_core/crypto.d.ts",
+      "import": "./dist/_core/crypto.js",
+      "require": "./dist/_core/crypto.cjs"
+    },
+    "./shared": {
+      "types": "./dist/_core/shared.d.ts",
+      "import": "./dist/_core/shared.js",
+      "require": "./dist/_core/shared.cjs"
+    },
     "./types": "./dist/@types/index.d.ts"
   },
   "keywords": [],
@@ -64,7 +79,9 @@
     "@aura-stack/tsconfig": "workspace:*",
     "@aura-stack/tsup-config": "workspace:*",
     "@testing-library/dom": "^10.4.1",
-    "@testing-library/react": "^16.3.2"
+    "@testing-library/react": "^16.3.2",
+    "react": "catalog:react",
+    "@types/react": "catalog:react"
   },
   "peerDependencies": {
     "react": ">=19.0.0",
