+++ packages/react/src/context.tsx
@@ -1,16 +1,7 @@
 "use client"
-
 import { createContext, useCallback, useEffect, useMemo, useState, useTransition } from "react"
-import type {
-    CredentialsPayload,
-    DeepPartial,
-    LiteralUnion,
-    BuiltInOAuthProvider,
-    Session,
-    SignInOptions,
-    SignOutOptions,
-    User,
-} from "@aura-stack/auth"
+import type { BuiltInOAuthProvider, Session, User } from "@aura-stack/auth"
+import type { DeepPartial, LiteralUnion, CredentialsPayload, SignInOptions, SignOutOptions } from "@aura-stack/auth/types"
 import type { AuthProviderProps, AuthReactContextValue, UpdateSessionCallOptions } from "@/@types/types.ts"
 
 /**
