+++ packages/react/src/index.tsx
@@ -1,5 +1,5 @@
+export { createAuth } from "@aura-stack/auth"
 export { createAuthClient, type AuthClientOptions } from "@aura-stack/auth/client"
 export { AuthProvider, type AuthProviderProps } from "@/context.tsx"
 export { useAuth, useSession, useSignIn, useSignInCredentials, useSignOut, useUpdateSession } from "@/hooks.ts"
-export { builtInOAuthProviders, type BuiltInOAuthProvider } from "@aura-stack/auth/oauth/index"
-export type { User, Session, AuthConfig, AuthInstance } from "@/@types/core.ts"
+export type { User, Session, AuthConfig, AuthInstance } from "@aura-stack/auth"
