+++ packages/react/tsup.config.ts
@@ -1,9 +1,12 @@
 import { defineConfig } from "tsup"
 import { tsupConfig } from "@aura-stack/tsup-config"
 
-export default defineConfig({
-    ...tsupConfig,
-    banner: {
-        js: `"use client"`,
+export default defineConfig([
+    {
+        ...tsupConfig,
+        entry: ["src", "!src/index.ts"],
+        banner: {
+            //js: `"use client"`,
+        },
     },
-})
+])
